kprobes: Fix kprobe selftest for gcc 4.6
Patchwork-id: 49252
O-Subject: [PATCH RHEL6.4 BZ786489 61/70] kprobes/trace: Fix kprobe selftest for gcc 4.6
Bugzilla: 786489
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 786489
https://bugzilla.redhat.com/show_bug.cgi?id=786489

upstream status
===============
commit 265a5b7ee3eb21a4d0e53e17d59ba6eada91af39
Author: Steven Rostedt <srostedt@redhat.com>
Date:   Mon Jun 6 22:35:13 2011 -0400

description
===========
With gcc 4.6, the self test kprobe function:

 kprobe_trace_selftest_target()

is optimized such that kallsyms does not list it. The kprobes
test uses this function to insert a probe and test it. But
it will fail the test if the function is not listed in kallsyms.

Adding a __used annotation keeps the symbol in the kallsyms table.

Signed-off-by: Jarod Wilson <jarod@redhat.com>