isci: make function declaration match implementation
Patchwork-id: 49279
O-Subject: [RHEL6.4 PATCH 8/12] isci: make function declaration match implementation
Bugzilla: 839070
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>
RH-Acked-by: Jeff Garzik <jgarzik@redhat.com>

http://marc.info/?l=linux-scsi&m=134204323014862&w=2

From: Dan Carpenter <dan.carpenter@oracle.com>

Sparse complains that we redeclare this with a different type, because
in the .c file we use an enum and in the .h file we declare the
parameter as a u32.  Probably it's best to use an enum in both places.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>

BZ 839070

Signed-off-by: Jarod Wilson <jarod@redhat.com>