tcp: clear hints to avoid a stale one
Patchwork-id: 49462
O-Subject: [PATCH RHEL6.4] tcp: clear hints to avoid a stale one
Bugzilla: 807704
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Several customers are reporting problems with RHEL6 that appear to be
caused by stale skb_hints in the tcp socket structure.  A patch to clear
these hints has been upstream since 2.6.33 and it looks like it will
resolve the customer's problem.

This is a backport of the following upstream commit:

	commit 8818a9d884e3a589899be3303958fff182e98e55
	Author: Ilpo Järvinen <ilpo.jarvinen@helsinki.fi>
	Date:   Wed Dec 2 22:24:02 2009 -0800

	    tcp: clear hints to avoid a stale one (nfs only affected?)

We have been unable to reproduce and cannot get any customers to try our
test kernels, but this should resolve RHBZ 807704.

Signed-off-by: Jarod Wilson <jarod@redhat.com>