raid1: Allocate spare to store replacement devices and their bios.
Patchwork-id: 49604
O-Subject: [RHEL6.4 PATCH 038/147] md/raid1: Allocate spare to store replacement devices and their bios.
Bugzilla: 817970
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: NeilBrown <neilb@suse.de>

In RAID1, a replacement is much like a normal device, so we just
double the size of the relevant arrays and look at all possible
devices for reads and writes.

This means that the array looks like it is now double the size in some
way - we need to be careful about that.
In particular, we checking if the array is still degraded while
creating a recovery request we need to only consider the first 'half'
- i.e. the real (non-replacement) devices.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 8f19ccb2fd70deb1f278be5e75076074cfddee46)

Conflicts:
	drivers/md/raid1.c

Minor fixup due to new plugging code not being present in RHEL6

Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>