bitmap: record the space available for the bitmap in the superblock.
Patchwork-id: 49688
O-Subject: [RHEL6.4 PATCH 131/147] md/bitmap: record the space available for the bitmap in the superblock.
Bugzilla: 817970
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: NeilBrown <neilb@suse.de>

Now that bitmaps can grow and shrink it is best if we record
how much space is available.  This means that when
we reduce the size of the bitmap we won't "lose" the space
for late when we might want to increase the size of the bitmap
again.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 1dff2b87a34a1ac1d1898ea109bf97ed396aca53)

Conflicts:
	drivers/md/bitmap.c

Caused by differences in kmap_atomic() API

Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>