driver-core: fix device_register race
Patchwork-id: 49718
O-Subject: [RHEL6.4 PATCH 1/1] driver-core: fix device_register race
Bugzilla: 833098
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Sebastian Ott <sebott@linux.vnet.ibm.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=833098
Upstream Status: accepted in scsi
Tested: tested by Netapp Engenio
Brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=4742921

driver-core: fix race between device_register and driver_register

When a device is registered to a bus it will be a) added to the list
of devices of the bus and b) bind to a driver (if one matches). As a
result of a driver being registered at this bus between a) and b) this
device could already be bound to a driver. This leads to a warning
and incorrect refcounting.
To fix this add a check to device_attach to identify an already bound
device.

Signed-off-by: Sebastian Ott <sebott@linux.vnet.ibm.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Jarod Wilson <jarod@redhat.com>