gfs2: Make gfs2_write_end not dirty the inode with every write
Patchwork-id: 49774
O-Subject: [RHEL6.4 PATCH] GFS2: Make gfs2_write_end not dirty the inode with every write
Bugzilla: 844814
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

Hi,

This patch is for bz #844814 - GFS2: Performance regression found when running Tibco EMS

Before this patch, function gfs2_write_end was marking the inode
dirty with every write. If the inode was opened as O_DSYNC, it
was doing a lot more writes than it needed to. This patch does
two things: First, it only marks the inode dirty if the inode
size has changed, just like generic_write_end does. Second, it
inlines a bunch of code from generic_write_end in order to
avoid a redundant dirtying of the inode which we were doing in
order to write out the dinode size after gfs2_dinode_out.

It was tested using the Tibco EMS app.

Without the patch:
50000 times took 185.012 seconds, performance is 270 messages/second

With the patch:
50000 times took 35.538 seconds, performance is 1406 messages/second

Note: This patch is NOT upstream and it doesn't need to be.
That's because new callbacks were added upstream for s_op->dirty_inode
that caused the affected code to be restructured, and the
restructuring made the performance regression go away. However,
those patches are too invasive to back-port to RHEL6, so this
is the next best thing.

rhbz#844814

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>