events: Add files to create libtraceevent.a
Patchwork-id: 49797
O-Subject: [PATCH RHEL6.4 BZ846702 001/198] tools/events: Add files to create libtraceevent.a
Bugzilla: 846702
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 846702
https://bugzilla.redhat.com/show_bug.cgi?id=846702

upstream status
===============
commit f7d82350e597d76dc8765a55c7849843395728b0
Author: Steven Rostedt <srostedt@redhat.com>
Date:   Fri Apr 6 00:47:53 2012 +0200

description
===========
Copy over the files from trace-cmd to the Linux tools directory
such that applications like perf and latencytrace can use the
more advanced parsing code.

Because some of the file names of perf conflict with trace-cmd file
names, the trace-cmd files have been renamed as follows:

 parse-events.c ==> event-parse.c
 parse-events.h ==> event-parse.h
 utils.h        ==> event-utils.h

The files have been updated to handle the changes to the header files
but other than that, they are identical to what was in the trace-cmd
repository. The history of these files, including authorship is
available at the git repo:

 git://git.kernel.org/pub/scm/linux/kernel/git/rostedt/trace-cmd.git

The Makefile was also copied over, but most of it was removed to
focus on the parse-events code first. The parts of the Makefile for
the plugins have also been removed, but will be added back when the
plugin code is copied over as well. But that may be in its own
separate directory.

Signed-off-by: Jarod Wilson <jarod@redhat.com>